BACKEND ::: New DB schema, implement logic from endpoint to ddbb for "unidad medida" and "clave presupuestal", CORS, Auditorio "fields" isolated from tables

+++ almacen-backend/src/main/java/com/sesesp/almacen/contrato/dto/ContratoRequest.java
@@ -7,10 +7,10 @@ public class ContratoRequest {
 
     private String identificadorContrato;
 
-    private Long idProveedor;
-    private Long idComprador;
-    private Long idAdministradorContrato;
-    private Long idTipoOrigenContrato;
+    private Integer idProveedor;
+    private Integer idComprador;
+    private Integer idAdministradorContrato;
+    private Integer idTipoOrigenContrato;
 
     private String folioOrigen;
     private LocalDateTime fechaOrigen;
@@ -30,7 +30,7 @@ public class ContratoRequest {
     private Double montoFiniquito;
 
     private LocalDateTime fechaTentativaLlegada;
-    private Long idEstatusContrato;
+    private Integer idEstatusContrato;
 
     public String getIdentificadorContrato() {
         return identificadorContrato;
@@ -40,35 +40,35 @@ public class ContratoRequest {
         this.identificadorContrato = identificadorContrato;
     }
 
-    public Long getIdProveedor() {
+    public Integer getIdProveedor() {
         return idProveedor;
     }
 
-    public void setIdProveedor(Long idProveedor) {
+    public void setIdProveedor(Integer idProveedor) {
         this.idProveedor = idProveedor;
     }
 
-    public Long getIdComprador() {
+    public Integer getIdComprador() {
         return idComprador;
     }
 
-    public void setIdComprador(Long idComprador) {
+    public void setIdComprador(Integer idComprador) {
         this.idComprador = idComprador;
     }
 
-    public Long getIdAdministradorContrato() {
+    public Integer getIdAdministradorContrato() {
         return idAdministradorContrato;
     }
 
-    public void setIdAdministradorContrato(Long idAdministradorContrato) {
+    public void setIdAdministradorContrato(Integer idAdministradorContrato) {
         this.idAdministradorContrato = idAdministradorContrato;
     }
 
-    public Long getIdTipoOrigenContrato() {
+    public Integer getIdTipoOrigenContrato() {
         return idTipoOrigenContrato;
     }
 
-    public void setIdTipoOrigenContrato(Long idTipoOrigenContrato) {
+    public void setIdTipoOrigenContrato(Integer idTipoOrigenContrato) {
         this.idTipoOrigenContrato = idTipoOrigenContrato;
     }
 
@@ -176,11 +176,11 @@ public class ContratoRequest {
         this.fechaTentativaLlegada = fechaTentativaLlegada;
     }
 
-    public Long getIdEstatusContrato() {
+    public Integer getIdEstatusContrato() {
         return idEstatusContrato;
     }
 
-    public void setIdEstatusContrato(Long idEstatusContrato) {
+    public void setIdEstatusContrato(Integer idEstatusContrato) {
         this.idEstatusContrato = idEstatusContrato;
     }
 }
